feat: AdTracker 분리 및 정리
